+++ Validetux/AbstractValidator.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Validetux.Models;
 
 namespace Validetux
 {
