+++ Validetux/Models/ParameterDetails.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Validetux
+namespace Validetux.Models
 {
     public class ParameterDetails<M>
     {
