+++ Validetux/Models/ValidationResult.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
-namespace Validetux
+namespace Validetux.Models
 {
     public class ValidationResult
     {
