+++ Validetux/RulesContext.cs
@@ -1,6 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using Validetux.Abstractions;
+using Validetux.Models;
 
 namespace Validetux
 {
